Adjust and portion the flexbox divisions
Adjust the targetting mechanism, switched from ID
to classes so that the same line of code will not
have to be applied to every single section.

Change the overall flexbox container to have a
downward direction to build a top down webpage
design.

Portion the sizes of the total screen.
Take the dimensions from the Odin Project
project reference image and change the dimensions
to accomodate a full 1920 pixel display while
keeping the same ratio.

Portion the main flexbox container in accordance
with the reference image, largest division at the
top and the divisions gradually decrease in height
as the page goes down.